Refactor: Organize UI component files
