+++ src/components/ui/common/command.jsx
@@ -4,7 +4,7 @@ import { Command as CommandPrimitive } from 'cmdk';
 import { Search } from 'lucide-react';
 
 import { cn } from '@/lib/utils';
-import { Dialog, DialogContent } from '@/components/ui/dialog';
+import { Dialog, DialogContent } from '@/components/ui/common/dialog';
 
 const Command = React.forwardRef(({ className, ...props }, ref) => (
   <CommandPrimitive
