+++ src/components/ui/tasks/data-table-faceted-filter.jsx
@@ -5,8 +5,8 @@ import { CheckIcon, PlusCircledIcon } from '@radix-ui/react-icons';
 import { Column } from '@tanstack/react-table';
 
 import { cn } from '@/lib/utils';
-import { Badge } from '@/components/ui/badge';
-import { Button } from '@/components/ui/button';
+import { Badge } from '@/components/ui/common/badge';
+import { Button } from '@/components/ui/common/button';
 import {
   Command,
   CommandEmpty,
@@ -15,13 +15,13 @@ import {
   CommandItem,
   CommandList,
   CommandSeparator,
-} from '@/components/ui/command';
+} from '@/components/ui/common/command';
 import {
   Popover,
   PopoverContent,
   PopoverTrigger,
-} from '@/components/ui/popover';
-import { Separator } from '@/components/ui/separator';
+} from '@/components/ui/common/popover';
+import { Separator } from '@/components/ui/common/separator';
 
 export function DataTableFacetedFilter({ column, title, options }) {
   const facets = column?.getFacetedUniqueValues();
