+++ src/components/ui/tasks/data-table-pagination.jsx
@@ -8,14 +8,14 @@ import {
 } from '@radix-ui/react-icons';
 import { Table } from '@tanstack/react-table';
 
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/ui/common/button';
 import {
   Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from '@/components/ui/select';
+} from '@/components/ui/common/select';
 
 export function DataTablePagination({ table }) {
   return (
